Making cookie id column 64 length so we dont exceed the mysql index size limit. Fixing error where all nulled cookie_id columns were getting updated on every request.

+++ src/Trackers/RequestTracker.php
@@ -55,10 +55,11 @@ class RequestTracker extends TrackerBase
 
         self::$lastTrackedRequestId = $requestId;
 
-        if ($userId) {
+        if (!empty($userId) && !empty($cookieId)) {
             $this->setUserIdOnOldRequests($userId, $cookieId);
             $this->deleteCookieForAuthenticatedUser();
         }
+
         return $requestId;
     }
